Ensure movie autostart works with pdfTeX (fixes issue #188)
Thanks to Alexander Grahn for the analysis here. Essentially,
we should be using \pdfpageattr, but there are some odd things
going on. Rather than have to work through them, and risk breaking
other stuff, a separate annotation is used to control the autostart.
This should work with players that work with the original
inclusion method.