Detect blank line read as \par when first line of verbatim frame (fixes issue #215)
When reading verbatim, we have an issue as beamer is rather too
flexible with the first token in the frame: see issue #93 for when this
can go wrong. If the first token in the frame is a blank line, TeX will
convert this to a \par token. The best way to deal with this is to
detect such a token and write as a blank line to the .vrb file. (It is
much more likely that a blank line has been read than an actual \par
token.)