Disable keyval errors for all options (see issue #236)
Till's code only deliberately targetted the class options but 'escaped'.
It's likely that the real target was other options, but as things stand
tightening up here will break code that relies on the old (incorrect)
approach. The best plan is to disable errors for all beamer options, but
not more generally (which is what was happening).